// src/app/models/Akun.ts
export class Akun {
  id_transaksi_akun: number;
  nama_transaksi_akun: string;
  debit_transaksi_akun: number;
  kredit_transaksi_akun: number;
  saldo_akun: number;
  status_debit_kredit: number;
  kelompok_hitungan: number;
  debit_nutup_harian: number;
  kredit_nutup_harian: number;
  saldo_nutup_harian: number;
  sts: string;
}